#222: fold BLELed into Led, rebuild tests/Led's variant matrix
Led and BLELed had drifted into two different feature sets by accident, not
design: BLELed started life as a copy from a separate standalone repo
(iotsaBLELed, 2019-12-09) and was simplified to static-color-only along the
way, never resynced with Led's own flash-pattern work (color/onDuration/
offDuration/repeat count). Merged IotsaLedControlMod now has both: full
pattern control over web/REST/CoAP, plus BLELed's BLE characteristic
(solid-color-only over BLE, by design -- three extra parameters don't map
cleanly onto one GATT characteristic) and its battery/sleep reference wiring
(VBAT/VUSB, disable-sleep pin, BLE-triggered config-mode switch).

tests/Led's variant matrix rebuilt from what was 16 accumulated (and partly
duplicate) rows down to 12 deliberate ones, organized around two axes instead
of one-flag-at-a-time:
- `all` (every optional transport a board supports, together) per board:
  nodemcuv2 (CoAP+HTTPS, no BLE -- unsupported on ESP8266), esp32thing,
  esp32c3devkit, esp32s3supermini, lolin32 (all three). Replaces the old
  withhttps/withcoap/ble-alone rows, which each toggled one flag against an
  otherwise-default build examples/Led already covers.
- Three "only" isolations, each a real distinct device shape rather than an
  arbitrary combinatorial point: `onlycoap` (CoAP as the sole protocol, zero
  HTTP surface -- constrained-network/gateway integration) and `onlyhttps`
  (HTTPS with no plain-HTTP fallback -- previously untested anywhere in the
  whole repo) on nodemcuv2+esp32thing; `onlyble` (BLE with WiFi entirely off,
  renamed from bleonly) on esp32s3supermini, since that's this example's
  actual stated purpose (a battery device with no permanent WiFi connection).
- `nonetworking` (nodemcuv2+esp32thing): every transport off, BLE included --
  reframes what were two accidental "nowifi" rows (redundant subsets of
  onlyble's WiFi-off code path) into one deliberate "total isolation, does it
  still boot" baseline.
- Cut: `lolin32/coap` (BLE+CoAP without REST/HTTPS) sat between `all` and
  `onlycoap` without clearly testing something neither brackets. Dropped a
  vestigial `-DIOTSA_WITH_PLACEHOLDERS` from onlycoap (only affects
  IotsaFiles*/IotsaWifiMod, none of which Led instantiates).

Since `nonetworking` has no network reachability at all, added an opt-in
startup blink (-DIOTSA_STARTUP_BLINK_COUNT=<n>, wired to the existing
set()/loop() flash-pattern machinery) as the only way to visually confirm
such a build actually booted; enabled on both nonetworking variants.

Verified: all 16 Led-related envs (4 example + 12 test) build green,
including lolin32-test-Led-all (BLE+CoAP+HTTPS) with no partition override
needed despite the README's "lolin32+BLE needs no_ota.csv" note.

Also updated the toplevel platformio.ini's default_envs, which still pointed
at the now-gone esp32thing-example-BLELed.
