--- examples/BLELed/BLELed.ino
@@ -1,1 +1,0 @@
-#include <iotsa.h>
