--- examples/BLELed/iotsa-build.json
@@ -1,35 +1,0 @@
-{
-  "defaults": {
-    "build_flags": [
-      "-DIOTSA_WITH_BLE"
-    ],
-    "partitions": "min_spiffs.csv",
-    "ci": {
-      "platformio": true,
-      "arduino": false
-    }
-  },
-  "variants": [
-    {
-      "board": "esp32thing"
-    },
-    {
-      "board": "esp32c3devkit",
-      "build_flags": [
-        "-DIOTSA_WITH_BLE",
-        "-DIOTSA_PIN_NEOPIXEL=8"
-      ]
-    },
-    {
-      "board": "esp32s3supermini",
-      "build_flags": [
-        "-DIOTSA_WITH_BLE",
-        "-DIOTSA_PIN_NEOPIXEL=48"
-      ],
-      "ci": {
-        "platformio": true,
-        "arduino": true
-      }
-    }
-  ]
-}
