--- examples/BLELed/platformio.ini
@@ -1,41 +1,0 @@
-; PlatformIO Project Configuration File
-; Generated by extras/python/gen_build_matrix.py -- do not hand-edit,
-; regenerate instead:
-;   python3 extras/python/gen_build_matrix.py --format=standalone-ini \
-;     --dir=examples/BLELed -o examples/BLELed/platformio.ini
-; To pin this example against a specific published iotsa branch/tag
-; instead of the local checkout, add an iotsa-config.json here:
-;   {"iotsa": "https://github.com/cwi-dis/iotsa.git#develop"}
-;
-[platformio]
-src_dir = .
-default_envs = esp32thing
-
-[common]
-framework = arduino
-lib_ldf_mode = deep+
-lib_compat_mode = strict
-lib_deps = symlink://../..
-monitor_speed = 115200
-
-[env:esp32thing]
-extends = common
-platform = espressif32
-board = esp32thing
-build_flags = -DIOTSA_WITH_BLE
-board_build.partitions = min_spiffs.csv
-
-[env:esp32c3devkit]
-extends = common
-platform = espressif32
-board = esp32-c3-devkitm-1
-build_flags = -DIOTSA_WITH_BLE -DIOTSA_PIN_NEOPIXEL=8
-board_build.partitions = min_spiffs.csv
-
-[env:esp32s3supermini]
-extends = common
-platform = espressif32
-board = esp32-s3-devkitc-1
-build_flags = -DIOTSA_WITH_BLE -DIOTSA_PIN_NEOPIXEL=48
-board_build.partitions = min_spiffs.csv
-
