+++ examples/Led/platformio.ini
@@ -33,10 +33,88 @@ board_build.partitions = min_spiffs.csv
 extends = common
 platform = espressif32
 board = esp32-c3-devkitm-1
+board_build.partitions = min_spiffs.csv
 
 [env:esp32s3supermini]
 extends = common
 platform = espressif32
 board = esp32-s3-devkitc-1
 board_build.partitions = min_spiffs.csv
+
+[env:nodemcuv2-all]
+extends = common
+platform = espressif8266
+board = nodemcuv2
+build_flags = -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
+
+[env:esp32thing-all]
+extends = common
+platform = espressif32
+board = esp32thing
+build_flags = -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
+board_build.partitions = min_spiffs.csv
+
+[env:esp32c3devkit-all]
+extends = common
+platform = espressif32
+board = esp32-c3-devkitm-1
+build_flags = -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS -DIOTSA_PIN_NEOPIXEL=8
+board_build.partitions = min_spiffs.csv
+
+[env:esp32s3supermini-all]
+extends = common
+platform = espressif32
+board = esp32-s3-devkitc-1
+build_flags = -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS -DIOTSA_PIN_NEOPIXEL=48
+board_build.partitions = min_spiffs.csv
+
+[env:lolin32-all]
+extends = common
+platform = espressif32
+board = lolin32
+build_flags = -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
+
+[env:nodemcuv2-onlycoap]
+extends = common
+platform = espressif8266
+board = nodemcuv2
+build_flags = -DIOTSA_WITH_COAP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS
+
+[env:esp32thing-onlycoap]
+extends = common
+platform = espressif32
+board = esp32thing
+build_flags = -DIOTSA_WITH_COAP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS
+board_build.partitions = min_spiffs.csv
+
+[env:nodemcuv2-onlyhttps]
+extends = common
+platform = espressif8266
+board = nodemcuv2
+build_flags = -DIOTSA_WITHOUT_HTTP -DIOTSA_WITH_HTTPS
+
+[env:esp32thing-onlyhttps]
+extends = common
+platform = espressif32
+board = esp32thing
+build_flags = -DIOTSA_WITHOUT_HTTP -DIOTSA_WITH_HTTPS
+board_build.partitions = min_spiffs.csv
+
+[env:esp32s3supermini-onlyble]
+extends = common
+platform = espressif32
+board = esp32-s3-devkitc-1
+build_flags = -DIOTSA_WITH_BLE -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_PIN_NEOPIXEL=48
+
+[env:nodemcuv2-nonetworking]
+extends = common
+platform = espressif8266
+board = nodemcuv2
+build_flags = -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_API -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_STARTUP_BLINK_COUNT=10
+
+[env:esp32thing-nonetworking]
+extends = common
+platform = espressif32
+board = esp32thing
+build_flags = -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_API -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_STARTUP_BLINK_COUNT=10
