+++ generated_envs.ini
@@ -3,24 +3,6 @@
 ; regenerate instead:
 ;   python3 extras/python/gen_build_matrix.py --format=platformio-ini -o generated_envs.ini
 
-[env:esp32thing-example-BLELed]
-extends = esp32thing
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_BLE
-board_build.partitions = min_spiffs.csv
-
-[env:esp32c3devkit-example-BLELed]
-extends = esp32c3devkit
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32c3devkit.build_flags} -DIOTSA_WITH_BLE -DIOTSA_PIN_NEOPIXEL=8
-board_build.partitions = min_spiffs.csv
-
-[env:esp32s3supermini-example-BLELed]
-extends = esp32s3supermini
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32s3supermini.build_flags} -DIOTSA_WITH_BLE -DIOTSA_PIN_NEOPIXEL=48
-board_build.partitions = min_spiffs.csv
-
 [env:nodemcuv2-example-Button]
 extends = nodemcuv2
 build_src_filter = +<*> +<../examples/Button>
@@ -314,46 +296,6 @@ lib_deps =
     adafruit/DHT sensor library
     adafruit/Adafruit Unified Sensor
 
-[env:esp32thing-test-BLELed-withble]
-extends = esp32thing
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_BLE
-board_build.partitions = min_spiffs.csv
-
-[env:esp32c3devkit-test-BLELed-withble]
-extends = esp32c3devkit
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32c3devkit.build_flags} -DIOTSA_WITH_BLE -DIOTSA_PIN_NEOPIXEL=8
-board_build.partitions = min_spiffs.csv
-
-[env:nodemcuv2-test-BLELed-nowifi]
-extends = nodemcuv2
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_API -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP
-
-[env:esp32thing-test-BLELed-nowifi]
-extends = esp32thing
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32thing.build_flags} -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_API -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP
-
-[env:esp32s3supermini-test-BLELed-bleonly]
-extends = esp32s3supermini
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${esp32s3supermini.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_PIN_NEOPIXEL=48
-board_build.partitions = min_spiffs.csv
-
-[env:lolin32-test-BLELed-coap]
-extends = lolin32
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${lolin32.build_flags} -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITHOUT_REST -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP
-board_build.partitions = min_spiffs.csv
-
-[env:lolin32-test-BLELed-kitchensink]
-extends = lolin32
-build_src_filter = +<*> +<../examples/BLELed>
-build_flags = ${lolin32.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
-board_build.partitions = min_spiffs.csv
-
 [env:nodemcuv2-test-KitchenSink]
 extends = nodemcuv2
 build_src_filter = +<*> +<../tests/KitchenSink>
@@ -372,38 +314,70 @@ build_src_filter = +<*> +<../tests/KitchenSink>
 build_flags = ${esp32c3devkit.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS -DIOTSA_PIN_NEOPIXEL=8
 board_build.partitions = min_spiffs.csv
 
-[env:nodemcuv2-test-Led-withhttps]
+[env:nodemcuv2-test-Led-all]
 extends = nodemcuv2
 build_src_filter = +<*> +<../examples/Led>
-build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITH_HTTPS
+build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
 
-[env:esp32thing-test-Led-withhttps]
+[env:esp32thing-test-Led-all]
 extends = esp32thing
 build_src_filter = +<*> +<../examples/Led>
-build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_HTTPS
+build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
 board_build.partitions = min_spiffs.csv
 
-[env:nodemcuv2-test-Led-withcoap]
-extends = nodemcuv2
+[env:esp32c3devkit-test-Led-all]
+extends = esp32c3devkit
 build_src_filter = +<*> +<../examples/Led>
-build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITH_COAP
-
-[env:esp32thing-test-Led-withcoap]
-extends = esp32thing
-build_src_filter = +<*> +<../examples/Led>
-build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_COAP
+build_flags = ${esp32c3devkit.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS -DIOTSA_PIN_NEOPIXEL=8
 board_build.partitions = min_spiffs.csv
+
+[env:esp32s3supermini-test-Led-all]
+extends = esp32s3supermini
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${esp32s3supermini.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS -DIOTSA_PIN_NEOPIXEL=48
+board_build.partitions = min_spiffs.csv
+
+[env:lolin32-test-Led-all]
+extends = lolin32
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${lolin32.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITH_COAP -DIOTSA_WITH_HTTPS
 
 [env:nodemcuv2-test-Led-onlycoap]
 extends = nodemcuv2
 build_src_filter = +<*> +<../examples/Led>
-build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITH_COAP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITH_PLACEHOLDERS
+build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITH_COAP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS
 
 [env:esp32thing-test-Led-onlycoap]
 extends = esp32thing
 build_src_filter = +<*> +<../examples/Led>
-build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_COAP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITH_PLACEHOLDERS
+build_flags = ${esp32thing.build_flags} -DIOTSA_WITH_COAP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS
 board_build.partitions = min_spiffs.csv
+
+[env:nodemcuv2-test-Led-onlyhttps]
+extends = nodemcuv2
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITHOUT_HTTP -DIOTSA_WITH_HTTPS
+
+[env:esp32thing-test-Led-onlyhttps]
+extends = esp32thing
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${esp32thing.build_flags} -DIOTSA_WITHOUT_HTTP -DIOTSA_WITH_HTTPS
+board_build.partitions = min_spiffs.csv
+
+[env:esp32s3supermini-test-Led-onlyble]
+extends = esp32s3supermini
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${esp32s3supermini.build_flags} -DIOTSA_WITH_BLE -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_PIN_NEOPIXEL=48
+
+[env:nodemcuv2-test-Led-nonetworking]
+extends = nodemcuv2
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${nodemcuv2.build_flags} -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_API -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_STARTUP_BLINK_COUNT=10
+
+[env:esp32thing-test-Led-nonetworking]
+extends = esp32thing
+build_src_filter = +<*> +<../examples/Led>
+build_flags = ${esp32thing.build_flags} -DIOTSA_WITHOUT_WIFI -DIOTSA_WITHOUT_HTTP -DIOTSA_WITHOUT_HTTPS -DIOTSA_WITHOUT_WEB -DIOTSA_WITHOUT_API -DIOTSA_WITHOUT_REST -DIOTSA_WITHOUT_COAP -DIOTSA_STARTUP_BLINK_COUNT=10
 
 [env:esp32thing-sandbox-BLEClient]
 extends = esp32thing
