+++ platformio.ini
@@ -9,9 +9,10 @@
 ; http://docs.platformio.org/page/projectconf.html
 [platformio]
 default_envs =
-    esp32thing-example-BLELed
-; Per-example/test [env:...] sections are generated from examples/*/iotsa-build.json
-; and tests/*/iotsa-build.json -- do not hand-edit generated_envs.ini, regenerate it:
+    esp32thing-example-Led
+; Per-example/test/sandbox [env:...] sections are generated from
+; examples/*/iotsa-build.json, tests/*/iotsa-build.json, and
+; sandbox/*/iotsa-build.json -- do not hand-edit generated_envs.ini, regenerate it:
 ;   python3 extras/python/gen_build_matrix.py --format=platformio-ini -o generated_envs.ini
 extra_configs =
     generated_envs.ini
