--- tests/BLELed/iotsa-build.json
@@ -1,92 +1,0 @@
-{
-  "source": "examples/BLELed",
-  "defaults": {
-    "partitions": "min_spiffs.csv",
-    "ci": {
-      "platformio": true,
-      "arduino": false
-    }
-  },
-  "variants": [
-    {
-      "board": "esp32thing",
-      "label": "withble",
-      "build_flags": [
-        "-DIOTSA_WITH_BLE"
-      ],
-      "ci": {
-        "platformio": true,
-        "arduino": true
-      }
-    },
-    {
-      "board": "esp32c3devkit",
-      "label": "withble",
-      "build_flags": [
-        "-DIOTSA_WITH_BLE",
-        "-DIOTSA_PIN_NEOPIXEL=8"
-      ]
-    },
-    {
-      "board": "nodemcuv2",
-      "label": "nowifi",
-      "_partitions_note": "null overrides the file default deliberately: -DIOTSA_WITHOUT_WIFI makes OTA unreachable in this build regardless of partition scheme, so there's nothing for min_spiffs.csv to protect here -- see iotsa#40.",
-      "partitions": null,
-      "build_flags": [
-        "-DIOTSA_WITHOUT_WIFI",
-        "-DIOTSA_WITHOUT_HTTP",
-        "-DIOTSA_WITHOUT_WEB",
-        "-DIOTSA_WITHOUT_API",
-        "-DIOTSA_WITHOUT_REST",
-        "-DIOTSA_WITHOUT_COAP"
-      ]
-    },
-    {
-      "board": "esp32thing",
-      "label": "nowifi",
-      "_partitions_note": "null overrides the file default deliberately: -DIOTSA_WITHOUT_WIFI makes OTA unreachable in this build regardless of partition scheme, so there's nothing for min_spiffs.csv to protect here -- see iotsa#40.",
-      "partitions": null,
-      "build_flags": [
-        "-DIOTSA_WITHOUT_WIFI",
-        "-DIOTSA_WITHOUT_HTTP",
-        "-DIOTSA_WITHOUT_HTTPS",
-        "-DIOTSA_WITHOUT_WEB",
-        "-DIOTSA_WITHOUT_API",
-        "-DIOTSA_WITHOUT_REST",
-        "-DIOTSA_WITHOUT_COAP"
-      ]
-    },
-    {
-      "board": "esp32s3supermini",
-      "label": "bleonly",
-      "build_flags": [
-        "-DIOTSA_WITH_BLE",
-        "-DIOTSA_WITHOUT_WIFI",
-        "-DIOTSA_WITHOUT_HTTP",
-        "-DIOTSA_WITHOUT_WEB",
-        "-DIOTSA_WITHOUT_REST",
-        "-DIOTSA_WITHOUT_COAP",
-        "-DIOTSA_PIN_NEOPIXEL=48"
-      ]
-    },
-    {
-      "board": "lolin32",
-      "label": "coap",
-      "build_flags": [
-        "-DIOTSA_WITHOUT_HTTPS",
-        "-DIOTSA_WITHOUT_REST",
-        "-DIOTSA_WITH_BLE",
-        "-DIOTSA_WITH_COAP"
-      ]
-    },
-    {
-      "board": "lolin32",
-      "label": "kitchensink",
-      "build_flags": [
-        "-DIOTSA_WITH_BLE",
-        "-DIOTSA_WITH_COAP",
-        "-DIOTSA_WITH_HTTPS"
-      ]
-    }
-  ]
-}
